Allow verifiers to return a value directly (<> message)
Previously it was only possible to specify a
"return parameter" on a verifier.

With this commit, it is possible to return any kind of
value. In cmdhist, there will be an attribute linked
to a dummy parameter named /yamcs/cmd/returnValue.

An example use case for this, is when a command is
expected to return something that should be seen
immediately, like is common with RPC-style communication
flows.

A returnValue can be specified both on success or fail.

A returnValue can be specified in addition to a 'message'.